+++ cypress/integration/adminCanCreateArticle.feature.js
@@ -1,4 +1,3 @@
-import { ContextConsumer } from 'react-is'
 import sizes from '../support/index'
 import TestHelpers from '../support/testhelper'
 
@@ -13,15 +12,20 @@ describe('Admin is able to create an article ', () => {
         TestHelpers.sizeCase(size, selection)
       })
 
-      it('is expected to show article creation page', () => {
-        cy.get('[data-cy=article-title]').type('Social Care in London')
-        cy.get('[data-cy=article-teaser]').type('This is some teaser text')
-        cy.get('[data-cy=article-body]').type(
-          'Some longer text to see that the body of the article creation is working correctly'
-        )
-        cy.get('[data-cy=article-author]').type('Johnny Cage')
-        cy.get('[data-cy=article-date]').type('2021-08-24')
-        cy.get('[data-cy=article-submit]').click();
+      describe('successfully', () => {
+        before(() => {
+          cy.intercept('POST', '**/articles/create', {
+            message: 'Your article has successfully been created',
+          })
+        })
+        it('is expected to show article creation page', () => {
+          cy.get('[data-cy=article-title]').type('Social Care in London')
+          cy.get('[data-cy=article-body]').type(
+            'Some longer text to see that the body of the article creation is working correctly'
+          )
+
+          cy.get('[data-cy=article-submit]').click()
+        })
       })
     })
   })
