+++ src/modules/Articles.js
@@ -27,18 +27,19 @@ const Articles = {
   },
 
   async create(article) {
-    let params = { article: article };
+    let params = { article: article }
+    debugger
     try {
       let response = await axios.post('api/articles', params, {
         headers: headers,
-      });
-      Articles.index();
+      })
+
       store.dispatch({
         type: 'SET_SUBMIT',
         payload: { status: true, message: response.data.message },
-      });
+      })
     } catch (error) {
-      errorHandler(error);
+      errorHandler(error)
     }
   },
 
@@ -61,13 +62,12 @@ const Articles = {
   },
 }
 
-
 export const imageEncoder = (file) =>
   new Promise((resolve, reject) => {
-    const reader = new FileReader();
-    reader.readAsDataURL(file);
-    reader.onload = () => resolve(reader.result);
-    reader.onerror = (error) => reject(error);
-  });
+    const reader = new FileReader()
+    reader.readAsDataURL(file)
+    reader.onload = () => resolve(reader.result)
+    reader.onerror = (error) => reject(error)
+  })
 
 export default Articles
